First submit of algorithm suite
Algorithm is based on the Forward-Backward iterative approach to solving
optimal control problems using the Adjoint Method. It assumes the
controller is represented with a finite dimensional parameter vector.
Iteration uses a gradient descent with momentum at the moment, but may
move to nonlinear conjugate gradient in the future. This code does not
yet take into account constraints in the control. The way to enforce
constraints currently is via the dynamics.