Modified Example Problem 2 to be one analogous to a 2D missile guidance scenario
Modified Example 2 to be analogous to a simplified 2D missile guidance
scenario where the control can only be applied orthogonal to the
velocity vector. Interestingly, the formulation proved to be quite
successful in general and generated a range of paths, depending on the
time constraint, that made a lot of sense. If one requests short
timeframes, it generates a path similar to the Brachistochrone curve. If
one seeks longer flights, it takes a path that lofts the trajectory a
bit and comes in with a steep angle. Very interesting. I have also found
that using a higher order basis is superior to the lower in general and
really doesn't affect runtime much relative to more costly things, such
as time integration.